requirements.txt
astroid==2.5.6
attrs==20.3.0
autopep8==1.5.7
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
google-api-core==1.26.3
google-auth==1.30.0
google-cloud-vision==2.3.1
googleapis-common-protos==1.53.0
grpcio==1.37.1
idna==2.10
iniconfig==1.1.1
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.9
pluggy==0.13.1
proto-plus==1.18.1
protobuf==3.15.8
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pylint==2.8.2
pyparsing==2.4.7
pytest==6.2.3
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
toml==0.10.2
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
